김규원 - CompanyController 세션값 추가

+++ test26/src/main/java/ksmart/project/test26/CompanyController.java
@@ -2,6 +2,8 @@ package ksmart.project.test26;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -19,7 +21,10 @@ public class CompanyController {
 	
 	// 목록조회
 	@RequestMapping(value="/company/companyList", method = RequestMethod.GET)
-	public String movie(Model model) {
+	public String movie(Model model, HttpSession session) {
+		if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
 		List<Company> list = companyDao.selectCompanyList();
 		model.addAttribute("list",list);
 		return "company/companyList";	
@@ -27,28 +32,43 @@ public class CompanyController {
     
 	// 입력페이지 요청
 	@RequestMapping(value="/company/companyInsert", method = RequestMethod.GET)
-	public String companyAdd() {
+	public String companyAdd(HttpSession session) {
+		if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
        System.out.println("companyAdd 메소드 출력");
        return "company/companyInsert";
 	}
    
 	// 입력처리 요청
     @RequestMapping(value="/companyAdd", method = RequestMethod.POST)
-	public String companyAdd(Company company) {
+	public String companyAdd(Company company, HttpSession session) {
+    	// 세션이 없으면 로그인 페이지로 리다이렉트 
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
         companyDao.insertCompany(company);
         return "redirect:/company/companyList"; 
 	}
     
 	// 업데이트 처리요청
     @RequestMapping(value="/company/companyUpdate", method = RequestMethod.POST)
-    public String companyUpdate(Company company){
+    public String companyUpdate(Company company, HttpSession session){
+    	// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
     	companyDao.updateCompany(company);
     	return "redirect:/company/companyList";
 	}
     
  	// 업데이트 정보요청
     @RequestMapping(value="/company/companyUpdate", method = RequestMethod.GET)
-	public String companyUpdate(Model model, @RequestParam(value="companyId", required=true) int companyId){
+	public String companyUpdate(Model model, HttpSession session, @RequestParam(value="companyId", required=true) int companyId){
+    	// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
     	Company company= companyDao.getCompany(companyId);
     	model.addAttribute("company", company);
     	return "company/companyUpdate";
@@ -56,8 +76,12 @@ public class CompanyController {
 
     // 삭제
 	@RequestMapping(value="/company/companyDelete", method = RequestMethod.GET)
-	public String companyRemove(@RequestParam(value="companyId", required=true) int companyId) {
-        companyDao.deleteCompany(companyId);
+	public String companyRemove(HttpSession session, @RequestParam(value="companyId", required=true) int companyId) {
+		// 세션이 없으면 로그인 페이지로 리다이렉트
+    	if(session.getAttribute("loginMember") == null ) {
+			return "redirect:/log/login"; 
+		}
+		companyDao.deleteCompany(companyId);
         return "redirect:/company/companyList";
     } 
 }
